LPS-63951 Cast to int

+++ modules/apps/web-experience-management/asset/asset-browser-web/src/main/java/com/liferay/asset/browser/web/display/context/AssetBrowserDisplayContext.java
@@ -389,7 +389,7 @@ public class AssetBrowserDisplayContext {
 				getListable(), false, false);
 		}
 		else {
-			total = AssetEntryLocalServiceUtil.searchCount(
+			total = (int)AssetEntryLocalServiceUtil.searchCount(
 				themeDisplay.getCompanyId(), getFilterGroupIds(),
 				themeDisplay.getUserId(), assetRendererFactory.getClassName(),
 				getSubtypeSelectionId(), getKeywords(), isShowNonindexable(),
